fix format number pool v3
